requirements.txt
asn1crypto==0.24.0
bitmex==0.2.2
bitmex-ws==0.3.1
bravado==10.2.2
bravado-core==5.10.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
Flask==1.0.2
gevent==1.4.0
greenlet==0.4.15
idna==2.8
isort==4.3.10
itsdangerous==1.1.0
Jinja2==2.10
jsonref==0.2
jsonschema==2.6.0
kafka-python==1.4.4
locustio==0.9.0
MarkupSafe==1.1.0
monotonic==1.5
msgpack==0.6.1
msgpack-python==0.5.6
Naked==0.1.31
pyasn1==0.4.5
pycparser==2.19
pycryptodome==3.7.3
PyMySQL==0.9.3
PySocks==1.6.8
python-dateutil==2.7.5
pytz==2018.9
PyYAML==3.13
pyzmq==17.1.2
requests==2.21.0
rfc3987==1.3.8
rsa==4.0
sentry-sdk==0.7.4
shellescape==3.4.1
simplejson==3.16.0
six==1.12.0
strict-rfc3339==0.7
swagger-spec-validator==2.4.3
urllib3==1.24.1
webcolors==1.8.1
websocket-client==0.46.0
Werkzeug==0.14.1
